Move char and string assertion tests to a dedicated test class.
